--- client/src/utils/API.js
@@ -1,6 +1,0 @@
-import axios from "axios";
-
-export default axios.create({
-  baseURL: "https://api.spotify.com/v1/",
-  responseType: "json",
-});
